ref: Decouple symlink creation from the registry target
